Fix for JDK 1.4.
git-svn-id: http://svn.codehaus.org/xstream/branches/v-1.4.x@2359 9830eeb5-ddf4-0310-9ef7-f4b9a3e3227e

+++ xstream/src/java/com/thoughtworks/xstream/mapper/OuterClassMapper.java
@@ -60,7 +60,7 @@ public class OuterClassMapper extends MapperWrapper implements Caching {
             if (len == serialized.length()) {
                 idx = 0;
             } else if (serialized.length() > len + 1 && serialized.charAt(len) == '-') {
-                idx = Integer.valueOf(serialized.substring(len + 1));
+                idx = Integer.valueOf(serialized.substring(len + 1)).intValue();
             }
             if (idx >= 0) {
                 final String[] innerFieldNames = getInnerFieldNames(type);
